# server/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_module_api_modules_upload_post": {
        "properties": {
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_module_api_modules_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModuleImageRegenerateRequest": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "visual_state_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual State Key"
          }
        },
        "required": [
          "kind",
          "item_id"
        ],
        "title": "ModuleImageRegenerateRequest",
        "type": "object"
      },
      "ModuleRead": {
        "properties": {
          "clues": {
            "items": {},
            "title": "Clues",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "handouts": {
            "default": [],
            "items": {},
            "title": "Handouts",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "npcs": {
            "items": {},
            "title": "Npcs",
            "type": "array"
          },
          "rag_status": {
            "default": "",
            "title": "Rag Status",
            "type": "string"
          },
          "rule_system": {
            "title": "Rule System",
            "type": "string"
          },
          "scenes": {
            "items": {},
            "title": "Scenes",
            "type": "array"
          },
          "theme": {
            "title": "Theme",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "triggers": {
            "default": [],
            "items": {},
            "title": "Triggers",
            "type": "array"
          },
          "truth": {
            "default": "",
            "title": "Truth",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "world_setting": {
            "additionalProperties": true,
            "title": "World Setting",
            "type": "object"
          }
        },
        "required": [
          "id",
          "title",
          "rule_system",
          "description",
          "theme",
          "world_setting",
          "scenes",
          "npcs",
          "clues",
          "created_at",
          "updated_at"
        ],
        "title": "ModuleRead",
        "type": "object"
      },
      "ModuleWrite": {
        "description": "手动新建/编辑模组的结构化内容。",
        "properties": {
          "clues": {
            "default": [],
            "items": {},
            "title": "Clues",
            "type": "array"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "handouts": {
            "default": [],
            "items": {},
            "title": "Handouts",
            "type": "array"
          },
          "npcs": {
            "default": [],
            "items": {},
            "title": "Npcs",
            "type": "array"
          },
          "rule_system": {
            "default": "coc",
            "title": "Rule System",
            "type": "string"
          },
          "scenes": {
            "default": [],
            "items": {},
            "title": "Scenes",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "triggers": {
            "default": [],
            "items": {},
            "title": "Triggers",
            "type": "array"
          },
          "truth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Truth"
          },
          "world_setting": {
            "additionalProperties": true,
            "default": {},
            "title": "World Setting",
            "type": "object"
          }
        },
        "required": [
          "title"
        ],
        "title": "ModuleWrite",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Player-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "TRPG Player",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/modules": {
      "get": {
        "operationId": "list_modules_api_modules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModuleRead"
                  },
                  "title": "Response List Modules Api Modules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Modules",
        "tags": [
          "modules"
        ]
      },
      "post": {
        "description": "手动新建模组（结构化内容，非 PDF 解析）。",
        "operationId": "create_module_api_modules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Module",
        "tags": [
          "modules"
        ]
      }
    },
...
    "/api/modules/upload": {
      "post": {
        "description": "上传模组：同步做文件抽取与格式校验（错误立即可见），解析转后台任务。\n\n返回 {job_id}；前端轮询 GET /upload/status/{job_id} 展示进度并取最终结果。",
        "operationId": "upload_module_api_modules_upload_post",
        "parameters": [
          {
            "in": "query",
            "name": "rule_system",
            "required": false,
            "schema": {
              "default": "coc",
              "title": "Rule System",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_module_api_modules_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Module",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/upload/status/{job_id}": {
      "get": {
        "description": "轮询上传解析任务：{status: running|done|failed, stage, percent, detail, result}。",
        "operationId": "upload_status_api_modules_upload_status__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Status",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{module_id}": {
      "delete": {
        "operationId": "delete_module_api_modules__module_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Module",
        "tags": [
          "modules"
        ]
      },
      "get": {
        "operationId": "get_module_api_modules__module_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Module",
        "tags": [
          "modules"
        ]
      },
      "put": {
        "description": "整体编辑模组结构化内容。",
        "operationId": "update_module_api_modules__module_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Module",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{module_id}/images/regenerate": {
      "post": {
        "description": "图片文件缺失时重新生成，并回写 scenes/npcs/clues 中的图片 URL。",
        "operationId": "regenerate_module_image_api_modules__module_id__images_regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModuleImageRegenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate Module Image",
        "tags": [
          "modules"
        ]
      }
    },
    "/api/modules/{module_id}/rag/rebuild": {
      "post": {
        "description": "（重）建模组原文 RAG 索引：存量模组补建 / 失败后重试。\n\n立即置 rag_status=indexing 并返回，嵌入在后台进行；前端轮询模组的\nrag_status 观察 indexing → ready/failed。",
        "operationId": "rebuild_module_rag_api_modules__module_id__rag_rebuild_post",
        "parameters": [
          {
            "in": "path",
            "name": "module_id",
            "required": true,
            "schema": {
              "title": "Module Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Module Rag",
        "tags": [
          "modules"
        ]
      }
    },
...
  }
}